+++ AlphaFS/Network/Host Class/DriveConnection.cs
@@ -31,7 +31,7 @@ namespace Alphaleonis.Win32.Network
       #region Constructors
 
       /// <summary>Creates a temporary connection to a network resource. The function can redirect a local device to a network resource, using the current user credentials.</summary>
-      /// <param name="remoteName">The network resource to connect to. The string can be up to MAX_PATH characters in length.</param>
+      /// <param name="remoteName">The network resource to connect to. The string can be up to <see cref="Filesystem.NativeMethods.MaxPath"/> characters in length.</param>
       public DriveConnection(string remoteName)
       {
          Share = remoteName;
@@ -44,7 +44,7 @@ namespace Alphaleonis.Win32.Network
       }
 
       /// <summary>Creates a temporary connection to a network resource. The function can redirect a local device to a network resource, using a user name and password.</summary>
-      /// <param name="remoteName">The network resource to connect to. The string can be up to MAX_PATH characters in length.</param>
+      /// <param name="remoteName">The network resource to connect to. The string can be up to <see cref="Filesystem.NativeMethods.MaxPath"/> characters in length.</param>
       /// <param name="userName">
       ///   The user name for making the connection. If <paramref name="userName"/> is <see langword="null"/>, the function uses the default
       ///   user name. (The user context for the process provides the default user name)
@@ -69,7 +69,7 @@ namespace Alphaleonis.Win32.Network
       }
 
       /// <summary>Creates a temporary connection to a network resource. The function can redirect a local device to a network resource, <see cref="NetworkCredential"/> can be supplied.</summary>
-      /// <param name="remoteName">The network resource to connect to. The string can be up to MAX_PATH characters in length.</param>
+      /// <param name="remoteName">The network resource to connect to. The string can be up to <see cref="Filesystem.NativeMethods.MaxPath"/> characters in length.</param>
       /// <param name="credentials">An instance of <see cref="NetworkCredential"/> which provides credentials for password-based authentication schemes such as basic, digest, NTLM, and Kerberos authentication.</param>
       /// <param name="prompt"><see langword="true"/> always pops-up an authentication dialog box.</param>
       public DriveConnection(string remoteName, NetworkCredential credentials, bool prompt)